requirements.txt
data-designer
nvidia-nat[langchain]
openai>=1.40
pydantic>=2.6
python-dotenv
numpy
tavily-python
pyyaml
tqdm
